fix: upstream tag pollution
Some of the tags from upstream are used to tag release candidates or end
of maintence for a branch. This would lead to situations whereby
`os-downstream-tag.py` would create a tag based off of these unwanted
tags. To resolve this we can exclude those tags from the `git describe`
command.