Add os-cherry-pop.py
Proposes a list of downstream patches to cherry pick for a new release
branch. The list is generated based on the downstream patches in a previous
release, and accounts for patches that have been merged into the new
release. It also looks for newer versions of a patch on later branches.